Add multiboot crate and get some info from multiboot
In general, I don't want to have a bunch of crates, however this seems
to be a well-designed interface to a datastructure that I already wrote
parsing code for. I could write this again, but I don't think there's
any meaningful value in it above and beyond the crate for the moment.

That equation may change in the future, but for the moment we're going
to be okay with this.