Take a stab at a mutex
This is in a sub crate, mostly becuase it can totally be tested on Linux
as well as in-kernel. Having a sub crate allows me to develop this
seperately, run cargo test, and then use the code in Cardinal.

This is kinda what I was trying to do with submodules in nightingale,
but never found much success at it.

Obviously, I haven't actually imported this into the kernel and run it
there yet, but baby steps.