Leak kernel physical memory
For some reason, just making these linkscript-defined values statics and
taking their addresses doesn't work for some linker nonsense. Rust seems
to be actually generating GOT-relative relocations in this static
freestanding binary, which is moderately concerning and probably worth
investiagation, but for now making assembly stubs that return the base
and top values will do, and this does seem to succesfully leak the right
PhysicalRange for the kernel.